requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
Django==2.1.5
django-admin==1.3.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-excel-response2==2.0.8
django-rest-auth==0.9.3
django-six==1.0.4
djangorestframework==3.9.1
entrypoints==0.3
flake8==3.7.5
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
PyArabic==0.6.5
pycodestyle==2.5.0
pyflakes==2.1.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
screen==1.0.1
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
xlwt==1.3.0
